refactor: rename modules for Binance and Kraken use cases for consistency
