+++ app/src/application/exchange.rs
@@ -1,4 +1,4 @@
-pub mod binance;
+pub mod binance_use_cases;
 pub mod exchange_balances_routine;
-pub mod kraken;
+pub mod kraken_use_cases;
 pub mod use_cases;
