+++ app/src/application/exchange/binance_use_cases.rs
